v2.15.0: trace + profile completed -- Windows capture, jail from a profile, armasm64
TODO.trace-profile/07-10:

- retrace-profile capture on Windows: no preload, so the launch
  delegates to retrace-win-run (found next to the profiler, with
  retrace.dll); temp config via GetTempFileNameA; the profile
  and jail emission path is shared portable code. The recipe-34
  flow is one command on every platform now.
- retrace-profile jail <profile.json> [--inside d.json] [-o ...]:
  emit a jail config from an existing profile (the update-the-jail
  step of the upgrade story -- profile old, upgrade, profile new,
  diff, jail). prof_from_json is the inverse of prof_to_json
  (SSOT: both live in aggregate.c); jail emission extracted from
  the CLI to jail.c.
- access_add fix: a repeat access was credited to items[lo] after
  a bsearch break although the match sat at items[mid] -- any
  profile with 2+ paths mis-credited hits and classes to the
  wrong row since v2.12.0. Caught by the new round-trip tests
  (test_profile_jail.c: jail shape, declared allowlist, profile
  round trip, doc-vs-trace parity, degenerate inputs).
- armasm64 wrapper dialect (wrapper_arm64.asm): live hooks on the
  MSVC-arm64 legs; ASM_MARM64 wiring; HAVE_WRAPPERS covers all
  four arch/toolchain combos. The gas arm64 twin moves to the
  label-free immediate-index design like the x64 twins.
- Windows wrapper test un-gated on MSVC with a bounded TIMEOUT
  (the content assertion rides every leg; a trampoline loop now
  fails in 120s instead of hanging the job for an hour).

Bumps: version.h 2.14.0 -> 2.15.0, retrace_cli.c banner,
nix/debian/fedora packaging, CHANGELOG.md.

+++ src/backends/preload_elf/x86_64/arch_spec_bottom.c
@@ -64,7 +64,7 @@ struct WrapperSystemVFrame {
 	long real_rsp;
 };
 
-long retrace_as_call_real(const void *real_impl,
+intptr_t retrace_as_call_real(const void *real_impl,
 	const struct FuncParam params[],
 	int params_cnt)
 {
